+++ src/games/chess/pieces/king.rs
@@ -141,10 +141,10 @@ impl<const T: usize> Piece<T> for KingPiece<T> {
         &self,
         board: &Board<T>,
         from: BitBoard<T>,
-        from_bit: u32,
+        from_bit: u16,
         piece_type: usize,
-        team: u32,
-        mode: u32,
+        team: u16,
+        mode: u16,
         to: BitBoard<T>,
     ) -> BitBoard<T> {
         self.get_attack_lookup(board, piece_type).expect("Could not find move lookup table for king")[from_bit as usize][0]
@@ -156,8 +156,8 @@ impl<const T: usize> Piece<T> for KingPiece<T> {
         board: &Board<T>,
         from: BitBoard<T>,
         piece_type: usize,
-        team: u32,
-        mode: u32,
+        team: u16,
+        mode: u16,
     ) -> BitBoard<T> {
         let lookup = self.get_attack_lookup(board, piece_type);
         match lookup {
@@ -238,9 +238,9 @@ impl<const T: usize> Piece<T> for KingPiece<T> {
         actions: &mut Vec<Move>,
         board: &Board<T>,
         piece_type: usize,
-        from: u32,
-        team: u32,
-        mode: u32,
+        from: u16,
+        team: u16,
+        mode: u16,
     ) {
         let rows = board.state.rows;
         let cols = board.state.cols;
